feat: API docs for the DS details; fixed packaging
