+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/CreatorsFinder.scala
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets.rest
+package io.renku.knowledgegraph.datasets
 
+import Dataset.DatasetCreator
 import cats.MonadThrow
 import cats.data.NonEmptyList
 import cats.effect.Async
@@ -26,7 +27,6 @@ import io.circe.Decoder
 import io.renku.graph.model.Schemas._
 import io.renku.graph.model.datasets._
 import io.renku.graph.model.persons.{Affiliation, Email, Name => UserName}
-import io.renku.knowledgegraph.datasets.model.DatasetCreator
 import io.renku.triplesstore.SparqlQuery.Prefixes
 import io.renku.triplesstore._
 import org.typelevel.log4cats.Logger
