+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/DatasetId.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets.rest
+package io.renku.knowledgegraph.datasets
 
 import io.renku.graph.model.datasets
