+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/DatasetSearchResult.scala
@@ -16,8 +16,10 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets.rest
+package io.renku.knowledgegraph.datasets
 
+import Dataset.DatasetCreator
+import DatasetSearchResult.ProjectsCount
 import cats.syntax.all._
 import io.circe.literal._
 import io.circe.{Encoder, Json}
@@ -26,8 +28,6 @@ import io.renku.graph.model.datasets.{Date, DatePublished, Description, Identifi
 import io.renku.graph.model.{GitLabUrl, projects}
 import io.renku.http.rest.Links.{Href, Link, Rel, _links}
 import io.renku.json.JsonOps._
-import io.renku.knowledgegraph.datasets.model.DatasetCreator
-import io.renku.knowledgegraph.datasets.rest.DatasetSearchResult.ProjectsCount
 import io.renku.tinytypes.constraints.NonNegativeInt
 import io.renku.tinytypes.json.TinyTypeEncoders._
 import io.renku.tinytypes.{IntTinyType, TinyTypeFactory}
