+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/PartsFinder.scala
@@ -16,15 +16,15 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets.rest
+package io.renku.knowledgegraph.datasets
 
+import Dataset.DatasetPart
 import cats.MonadThrow
 import cats.effect.Async
 import eu.timepit.refined.auto._
 import io.circe.Decoder.decodeList
 import io.renku.graph.model.Schemas._
 import io.renku.graph.model.datasets._
-import io.renku.knowledgegraph.datasets.model.DatasetPart
 import io.renku.triplesstore.SparqlQuery.Prefixes
 import io.renku.triplesstore._
 import org.typelevel.log4cats.Logger
