+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/ProjectsFinder.scala
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets.rest
+package io.renku.knowledgegraph.datasets
 
+import Dataset.DatasetProject
 import cats.MonadThrow
 import cats.effect.kernel.Async
 import cats.syntax.all._
@@ -30,7 +31,6 @@ import io.renku.graph.model.datasets.Identifier
 import io.renku.graph.model.projects
 import io.renku.graph.model.projects.{Path, ResourceId, Visibility}
 import io.renku.http.server.security.model.AuthUser
-import io.renku.knowledgegraph.datasets.model.DatasetProject
 import io.renku.triplesstore.SparqlQuery.Prefixes
 import io.renku.triplesstore._
 import org.typelevel.log4cats.Logger
