+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/datasets/DatasetSearchEndpointDocsSpec.scala
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2022 Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.renku.knowledgegraph.datasets
+
+import io.renku.config.renku
+import io.renku.generators.CommonGraphGenerators.renkuApiUrls
+import io.renku.generators.Generators.Implicits._
+import io.renku.graph.model.GitLabUrl
+import io.renku.graph.model.GraphModelGenerators.gitLabUrls
+import io.renku.knowledgegraph.docs.OpenApiTester._
+import org.scalatest.wordspec.AnyWordSpec
+
+class DatasetSearchEndpointDocsSpec extends AnyWordSpec {
+
+  "path" should {
+
+    "return a valid Path object" in {
+      validatePath(new DatasetSearchEndpointDocsImpl().path)
+    }
+  }
+
+  private implicit lazy val renkuUrl:  renku.ApiUrl = renkuApiUrls.generateOne
+  private implicit lazy val gitLabUrl: GitLabUrl    = gitLabUrls.generateOne
+}
